requirements.txt
feedparser
requests
beautifulsoup4
google-genai
pytz
python-dotenv
streamlit